+++ tests/api/distances/test_comoving.py
@@ -38,6 +38,7 @@ def test_compliant_comoving(cosmology_cls):
         bases=(cosmology_cls,),
         namespace={
             "comoving_distance": _return_1arg,
+            "inv_comoving_distance": _return_1arg,
             "transverse_comoving_distance": _return_1arg,
             "comoving_volume": _return_1arg,
             "differential_comoving_volume": _return_1arg,